refactor(vendo): wire route table — shared plumbing + threads (kill-list B4)
